requirements.txt
openai
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pytrends
python-dotenv
feedparser
markdown
pandas
